requirements.txt
# Core
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# API & HTTP
requests==2.31.0
python-dotenv==1.0.0

# Technical Indicators
ta==0.10.2

# Telegram Bot
python-telegram-bot==20.5

# Dashboard
streamlit==1.27.0
plotly==5.17.0
streamlit-autorefresh

# Scheduler
apscheduler==3.10.4

# Utilities
joblib==1.3.2
pytz==2023.3